requirement.txt
jupyterlab
python-dotenv
langchain-openai
openai